+++ src/captions/captions.ts
@@ -100,14 +100,18 @@ export namespace Captions {
     }
 
     /** Create captions for `validation` image type. */
-    export function forGeometryValidation(context: { entryId: string, view: ViewType }): ImageSpec {
-        const { entryId, view } = context;
+    export function forGeometryValidation(context: { entryId: string, view: ViewType, validationAvailable: boolean }): ImageSpec {
+        const { entryId, view, validationAvailable } = context;
         const description = new TextBuilder();
         description.push('The deposited structure of PDB entry', B_, entryId, _B, 'coloured by geometry validation', ',', viewPhrase(view), '.');
-        description.push('Residues are coloured by the number of geometry outliers: green – no outliers, yellow – one outlier yellow, orange – two outliers, red – three or more outliers.');
+        if (validationAvailable) {
+            description.push('Residues are coloured by the number of geometry outliers: green – no outliers, yellow – one outlier yellow, orange – two outliers, red – three or more outliers.');
+        } else {
+            description.push('Validation data not available for this entry!');
+        }
         return {
             filename: `${entryId}_validation_geometry_deposited${viewSuffix(view)}`,
-            alt: new TextBuilder().push('Geometry outliers in PDB entry', entryId, ',', viewPhrase(view), '.').buildPlainText(),
+            alt: new TextBuilder().push('Geometry outliers in PDB entry', entryId, ',', viewPhrase(view), '-', validationAvailable ? '' : 'data not available', '.').buildPlainText(),
             description: description.buildText(),
             clean_description: description.buildPlainText(),
             _entry_id: entryId,