+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fixed bug in checking expected files (do not check if empty when missing)
+- Do not fail when PDBe Structure Validation Report is not available, show gray structure instead
 
 ## [2.1.0] - 2023-07-20